Update patch set 3
Patch Set 3:

(2 comments)

This is all good, but a thought did occur to me:

I'm coming from a position of complete ignorance here, but if this is just using the Amazon S3 API, is it possible to use the AWS packages to do this instead? There'd be some small changes to upspin-setupstorage-aws and cloud/storage/s3, but it should be pretty straightforward.

Patch-set: 3
CC: Gerrit User 5389 <5389@0759ed4b-3fdc-34dc-866b-f2e708168c03>